[IMP] Add a purchase requisition form view redefinition.
Is needed redefine the view two times; and the diference is that each
view add a diferent level of security.
The first add the security restriction for the group_purchase_user,
and avoid that user can modify a purchase requisition
fields (ej. product_id .. product_qty).
The second view definition add the security restriction for
group_purchase_manager who have true access to all the fields in the
form view.